requirements.txt
Jinja2>=2.10
MarkupSafe>=1.1.1
PyYAML>=5.1
